clear main.py, merge utillis, add pid in .gitignore
