+++ .gitignore
@@ -3,3 +3,4 @@
 */.venv/
 __pycache__/
 */*_cache/
+backend.pid
