--- backend/py/db/db_init.py
@@ -1,9 +1,0 @@
-def db_ensure_table(conn):
-    conn.execute("""
-        CREATE TABLE IF NOT EXISTS todos (
-            id SERIAL PRIMARY KEY,
-            text TEXT,
-            "dueDate" TEXT,
-            completed BOOLEAN DEFAULT FALSE
-        )
-    """)
